+++ package.json
@@ -26,6 +26,8 @@
     "@typescript-eslint/parser": "^5.55.0",
     "@vitejs/plugin-react": "^3.1.0",
     "axios": "^1.3.4",
+    "date-fns": "^2.29.3",
+    "dotenv": "^16.0.3",
     "eslint-plugin-react-hooks": "^4.6.0",
     "fetch-mock": "^9.11.0",
     "identity-obj-proxy": "^3.0.0",
