+++ src/app/routes/routes.ts
@@ -1,0 +1,8 @@
+const ROUTES = {
+  mainPage: '/',
+  aboutPage: '/about',
+  formPage: '/form',
+  errorPage: '/404',
+};
+
+export default ROUTES;
