+++ src/app/ui/atoms/input/input.tsx
@@ -3,16 +3,14 @@ import { InputProps } from './input.types';
 
 const Input = ({ className, placeholder, value, type = 'text', ...rest }: InputProps) => {
   return (
-    <>
-      <input
-        type={type}
-        className={className}
-        placeholder={placeholder}
-        data-testid="search-input"
-        value={value}
-        {...rest}
-      />
-    </>
+    <input
+      type={type}
+      className={className}
+      placeholder={placeholder}
+      data-testid="search-input"
+      value={value}
+      {...rest}
+    />
   );
 };
